### 2024-11-14 #7479 @editorialbot
[PRE REVIEW]: ReadmeReady: Free and Customizable Code Documentation with LLMs - A Fine-Tuning Approach
**Submitting author:** @souradipp76 (<a href="http://orcid.org/0000-0002-5781-3032">Souradip Pal</a>)
**Repository:** https://github.com/souradipp76/ReadMeReady
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.0
