+++ paper/paper.bib
@@ -285,7 +285,7 @@
 @inproceedings{
 hu2022lora,
 title={Lo{RA}: Low-Rank Adaptation of Large Language Models},
-author={Edward J Hu and Yelong Shen and Phillip Wallis and Zeyuan Allen-Zhu and Yuanzhi Li and Shean Wang and Lu Wang and Weizhu Chen},
+author={Edward J. Hu and Yelong Shen and Phillip Wallis and Zeyuan Allen-Zhu and Yuanzhi Li and Shean Wang and Lu Wang and Weizhu Chen},
 booktitle={International Conference on Learning Representations},
 year={2022},
 url={https://openreview.net/forum?id=nZeVKeeFYf9},
@@ -293,12 +293,12 @@ doi={10.48550/arXiv.2106.09685}
 }
 
 @article{zhang2023dynamically,
-  title={: A Dynamically Adaptive Defense to a Novel Attack on Review Fraud Detection Engines},
+  title={SockDef: A Dynamically Adaptive Defense to a Novel Attack on Review Fraud Detection Engines},
   author={Zhang, Youzhi and Chakrabarty, Sayak and Liu, Rui and Pugliese, Andrea and Subrahmanian, VS},
   journal={IEEE Transactions on Computational Social Systems},
   year={2023},
   publisher={IEEE},
-  doi={10.1109/TCSS.2023.3248149}
+  doi={10.1109/TCSS.2023.3321345}
 }
 
 @article{makarychev2024single,
