### 2024-11-14 #7479 @Manvi-Agrawal
Hi @crvernon , @souradipp76, I would like to review this paper. Can I go ahead?

**Brief about me:**
I am currently doing research on grammar inference using LLMs under Prof Miryung Kim at UCLA. I completed my masters in quantum from UCLA and undergrad in Computer Science from NSIT Delhi. Apart from this I have worked 2 years in Microsoft as a Software Engineer.
