### 2024-11-14 #7479 @camilochs
> 👋 @camilochs - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hello, @crvernon. Of course. I will review it during tomorrow.
