### 2024-11-17 souradipp76/ReadMeReady#8 @Manvi-Agrawal
[JOSS-Review][Question] How to use framework for non OpenAI models?
@souradipp76 , I am trying to install and use this python package. Could you please provide more details in the README on how to run a non OpenAI model since the main USP of this project is to have a framework for README generation to avoid costly OpenAI API pricing.
![image](https://github.com/user-attachments/assets/1fca0c8b-84dd-41e8-8a88-c2867bc8500f)

### Additional Context
As part of https://github.com/openjournals/joss-reviews/issues/7489
