### 2024-11-17 souradipp76/ReadMeReady#9 @Manvi-Agrawal
[JOSS-Review][Third-Party guidelines] Add guidelines on how to seek support
### JOSS Reviewer checklist item
**JOSS Review Issue:** https://github.com/openjournals/joss-reviews/issues/7489
Are there clear guidelines for third parties wishing to:
1. **Contribute to the software:** CONTRIBUTING.md
2. **Report issues or problems with the software** : README.md
3. **Seek support**: Currently, README/CONTRIBUTING doesnt state how should someone ask for help? Are there any dedicated slack channels/discord server? Mailing list? Should someone open github issue?

### Minor Nitpick[Non Blocking]
I would appreciate if the details about software issues could be moved to CONTRIBUTING.md. I have a habit to skim through the things. Consider Q# style, brief in README, expands in CONTRIBUTING: https://github.com/microsoft/qsharp?tab=readme-ov-file#contributing.
This is just a minor suggestion, not too binding, since this repository contains good issue, bug and PR templates.
