### 2024-11-17 souradipp76/ReadMeReady#10 @Manvi-Agrawal
[JOSS-Review][Question] What is meant by `application` in the Methodology section of the paper
Could you please clarify what is meant by `application` term used in the Methodology section of the paper?  Is there a website, app, etc which provides users to enter the project details?
> The application prompts the user to enter the project's name, GitHub URL, and select the desired model from the following options:

### Context
Part of https://github.com/openjournals/joss-reviews/issues/7489
