### 2024-11-17 souradipp76/ReadMeReady#11 @Manvi-Agrawal
[JOSS-Review][Reproducibility] How to reproduce the results?
How to verify the claims made in the paper? Could you please provide more details? For example: [this QuantumSE.jl](https://zenodo.org/records/10781381) artifact describes how to reproduce results in detail. It includes information regarding the docker image and the commmands to run inside the docker container and expected output(s).

### Additional Context
As part of https://github.com/openjournals/joss-reviews/issues/7489
