### 2024-11-17 souradipp76/ReadMeReady#12 @Manvi-Agrawal
[JOSS-Review][Question] Hardware Requirements?
What are the hardware requirements to run `readme-ready` framework?

### Additional Context
As part of https://github.com/openjournals/joss-reviews/issues/7489
