### 2024-11-17 souradipp76/ReadMeReady#13 @Manvi-Agrawal
[JOSS-Review] API Documentation
### JOSS Reviewer checklist item: See [this](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation)
**JOSS Review Issue:** https://github.com/openjournals/joss-reviews/issues/7489
- Consider more useful API documentation. Consider following [google style guide](https://google.github.io/styleguide/pyguide.html#38-comments-and-docstrings) or any other standard python guide.
- Consider generating a documentation website for API documentation. This makes the life of reviewers and API users easy. For example: see how [amazon-braket/amazon-braket-sdk-python](https://github.com/amazon-braket/amazon-braket-sdk-python) generates [API Reference](https://amazon-braket-sdk-python.readthedocs.io/en/latest/_apidoc/modules.html#). I saw the docstring for `generate_readme` function that is described in exammple usage section, and API documentation is not useful.
