### 2024-11-17 #7489 @Manvi-Agrawal
Hi @souradipp76, I did a high-level pass and left an initial review. I would appreciate more insights into how to reproduce the results and run `readme-ready` locally for non OpenAI models(I exhausted my free credits xD). After you provide with these details, I can do another round of review.
