+++ examples/example.py
@@ -10,7 +10,7 @@ from readme_ready.types import (
     LLMModels,
 )
 
-model = LLMModels.TINYLLAMA_1p1B_CHAT_GGUF # Choose model from supported models
+model = LLMModels.LLAMA2_7B_CHAT_HF # Choose model from supported models
 
 repo_config = AutodocRepoConfig (
     name = "readmy_ready",
@@ -41,7 +41,7 @@ repo_config = AutodocRepoConfig (
     priority = None,
     max_concurrent_calls = 50,
     add_questions = False,
-    device = "cpu",
+    device = "cuda",
 )
 
 user_config = AutodocUserConfig(
