### 2024-11-19 56a0a0a
Fixed device settings.
