+++ HISTORY.md
@@ -2,6 +2,342 @@ Changelog
 =========
 
 
-0.1.0 (2024-06-08)
-------------------
-- Initial Release.
+(unreleased)
+------------
+- Updated Makefile. [Souradip Pal]
+- Fixed device settings. [Souradip Pal]
+- Minor change. [Souradip Pal]
+- Fixed device map issue. [Souradip Pal]
+- Minor fixes. [Souradip Pal]
+- Merge pull request #14 from
+  souradipp76/dependabot/github_actions/codecov/codecov-action-5.
+  [Souradip Pal]
+
+  Bump codecov/codecov-action from 4 to 5
+- Bump codecov/codecov-action from 4 to 5. [dependabot[bot]]
+
+  Bumps [codecov/codecov-action](https://github.com/codecov/codecov-action) from 4 to 5.
+  - [Release notes](https://github.com/codecov/codecov-action/releases)
+  - [Changelog](https://github.com/codecov/codecov-action/blob/main/CHANGELOG.md)
+  - [Commits](https://github.com/codecov/codecov-action/compare/v4...v5)
+
+  ---
+  updated-dependencies:
+  - dependency-name: codecov/codecov-action
+    dependency-type: direct:production
+    update-type: version-update:semver-major
+  ...
+- Adjusted prompts. [Souradip Pal]
+- Fixed tests and updated CONTRIBUTING.md. [Souradip Pal]
+- Fixing utils. [Souradip Pal]
+- Minor fix. [Souradip Pal]
+- Updated example. [Souradip Pal]
+- Updated README and added example. [Souradip Pal]
+- Modified paper.md. [Souradip Pal]
+- Removed tokens. [Souradip Pal]
+- Update VERSION. [Sayak Chakrabarty]
+- Update main.yml. [Sayak Chakrabarty]
+- Rename project. [Souradip Pal]
+- Update setup.py. [Sayak Chakrabarty]
+- Update README.md. [Sayak Chakrabarty]
+- Update VERSION. [Sayak Chakrabarty]
+- Update main.yml. [Sayak Chakrabarty]
+- Updated README. [Souradip Pal]
+- Fixed tests and bib. [Souradip Pal]
+- Tests. [hellokayas]
+- Tests. [hellokayas]
+- Tests. [hellokayas]
+- Partially fixed tests. [Souradip Pal]
+- Tests. [hellokayas]
+- Merge branch 'main' of https://github.com/souradipp76/doc_generator.
+  [hellokayas]
+- Corrected qa mode and test. [Souradip Pal]
+- Fixed scripts and README. [Souradip Pal]
+- Added missing tests and fixed formatting. [Souradip Pal]
+- Tests. [hellokayas]
+- Fixed tests and workflow. [Souradip Pal]
+- Merge branch 'main' of https://github.com/souradipp76/ReadMeReady.
+  [Souradip Pal]
+- Tests. [hellokayas]
+- Merge branch 'main' of https://github.com/souradipp76/ReadMeReady.
+  [Souradip Pal]
+- Merge branch 'main' of https://github.com/souradipp76/doc_generator.
+  [hellokayas]
+- Tests. [hellokayas]
+- Updated workflow for win platform. [Souradip Pal]
+- Updated workflow. [Souradip Pal]
+- Modified workflow env. [Souradip Pal]
+- Added OPENAI_API_KEY in workflow env. [Souradip Pal]
+- Fixed tests. [Souradip Pal]
+- Test. [hellokayas]
+- Tests. [hellokayas]
+- Fixed utils test. [Souradip Pal]
+- Tests. [hellokayas]
+- Tests. [hellokayas]
+- Fixed linting errors. [Souradip Pal]
[…]
