### 2024-11-19 souradipp76/ReadMeReady#8 @souradipp76
Hi @Manvi-Agrawal, you can follow the installation and CLI usage instructions given in [README.md](https://github.com/souradipp76/ReadMeReady/blob/main/README.md#installation). For using non-OpenAI models, set a dummy value for OPENAI_API_KEY env variable and run the CLI app using the `readme_ready` command. The app will prompt you to select the open-source model and other parameters.
