### 2024-11-19 souradipp76/ReadMeReady#10 @souradipp76
Hi @Manvi-Agrawal, the project has a CLI application. You can follow the installation and CLI usage instructions given in [README.md](https://github.com/souradipp76/ReadMeReady/blob/main/README.md#installation) and run the CLI app using the `readme_ready` command. The app will prompt you to select the open-source model and other parameters.
