### 2024-11-19 souradipp76/ReadMeReady#15 @Manvi-Agrawal
[JOSS-Review] Installation + CLI Issues
### JOSS Reviewer checklist item: Installation
**JOSS Review Issue:** https://github.com/openjournals/joss-reviews/issues/7489

### Installation Issue
After doing `pip install readme-ready` and the running CLI command `readme-ready`, I am getting the following error:
`readme-ready: The term 'readme-ready' is not recognized as a name of a cmdlet, function, script file, or executable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.`

**System Info:**
OS: Windows 11
Python 3.11.9
pip: 24.3.1
