### 2024-11-20 souradipp76/ReadMeReady#15 @souradipp76
Hi @Manvi-Agrawal, the CLI command is `readme_ready` with an underscore to launch the CLI app.
You can also check out the code example at `examples/example.py`. You might have to change the configuration parameters based on your setup.
