### 2024-11-20 souradipp76/ReadMeReady#15 @Manvi-Agrawal
My bad. OOPS!
Now, I am getting : "ImportError: failed to find libmagic.  Check your installation". I installed python-libmagic-bin and this error got fixed.

 On entering these details in CLI, i get:

Initializing Auto Documentation...
? Project Name?[Example: readme_ready] readme_ready
? Project Root?[Example: ./readme_ready/] ./readme_ready/
? Project URL?[Example: https://github.com/username/readme_ready] https://github.com/souradipp76/ReadMeReady
? Output Directory?[Example: ./output/readme_ready/] ./output/readme_ready/
? Documentation Mode? Readme
? List of Readme Headings?(comma separated)[Example: # Introduction,## Usage]  # Introduction,## Usage
? Which model? meta-llama/Llama-2-7b-chat-hf
? Is finetuned? No
? Device? cpu
Initialization Complete.

Processing repository...
The provided folder path does not exist.
Processing 0 files and 0 folders...
The provided folder path does not exist.
Processing complete.
Creating markdown files...
Creating vector files...
[WinError 3] The system cannot find the path specified: './readme_ready/'
Traceback (most recent call last):
  File "C:\Users\manvi\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.11_qbz5n2kfra8p0\LocalCache\local-packages\Python311\site-packages\readme_ready\index\create_vector_store.py", line 52, in process_directory
    files = os.listdir(directory_path)
            ^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [WinError 3] The system cannot find the path specified: './readme_ready/'

Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "C:\...\readme_ready\__main__.py", line 6, in <module>
    main()
  File "C:\...\readme_ready\main.py", line 190, in main
    index.index(repo_conf)
  File "C:\...\readme_ready\index\index.py", line 75, in index
    create_vector_store(
  File "C:\...\readme_ready\index\create_vector_store.py", line 100, in create_vector_store
    raw_docs = loader.load()
               ^^^^^^^^^^^^^
  File "C:\...\readme_ready\index\create_vector_store.py", line 85, in load
    return process_directory(self.file_path, self.ignore)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "C:\...\readme_ready\index\create_vector_store.py", line 55, in process_directory
    raise FileNotFoundError(
FileNotFoundError: Could not read directory: ./readme_ready/.
