### 2024-11-20 souradipp76/ReadMeReady#15 @souradipp76
Project Root refers to the path to the code repository you want to generate a README for. Please download any sample code repository from GitHub and use that or if you have your own repository downloaded (say `MyRepo`) you can use that as well. Set Project Root as `<path to MyRepo>` and set the other parameters accordingly. Output Directory is where the README will be generated(you can set any appropriate path). For Windows platforms, please provide the paths correctly (e.g. C:\Users\_user_\Downloads\MyRepo).

Also, I would suggest using GPU since LLMs like `meta-llama/Llama-2-7b-chat-hf` take a lot of resources (~13.5 GB) and may run slowly on CPU, even with 4bit quantization(added by default).

Also, please upgrade the package to its latest version 1.1.3.
