### 2024-11-20 souradipp76/ReadMeReady#15 @souradipp76
As for the error "ImportError: failed to find libmagic", I will add an additional step for installation in the documentation to install the `python-magic`[link](https://pypi.org/project/python-magic/) since the installation is different for Windows platform.
