### 2024-11-20 souradipp76/ReadMeReady#15 @Manvi-Agrawal
@souradipp76 I don't have GPU access. Do you think T4 GPU provided by collab would work for verification?
