### 2024-11-20 souradipp76/ReadMeReady#15 @souradipp76
Yes. You can use T4 GPU provided by Google Colab for verification.
