### 2024-11-20 souradipp76/ReadMeReady#12 @souradipp76
@Manvi-Agrawal, the hardware requirements for the package depend on the open-source large language models used for text generation. Some of the models are large and resource-intensive so 4-bit quantization techniques have been applied on those models by default. For our experimentation and tests, we used 1 × NVIDIA Tesla V100 with 16GB of GPU memory. Thus a 16GB GPU is ideal for running the framework.
