### 2024-11-20 souradipp76/ReadMeReady#9 @souradipp76
@Manvi-Agrawal, I have added a section in [CONTRIBUTING.md](https://github.com/souradipp76/ReadMeReady/blob/main/CONTRIBUTING.md#questions-and-reporting-issues) instructing users to simply raise a GitHub issue in case of questions.
