### 2024-11-21 souradipp76/ReadMeReady#11 @souradipp76
Hi @Manvi-Agrawal, please find the details to reproduce the results under the 'Experiments and Results' section in the `full_paper.md`([link](https://github.com/souradipp76/ReadMeReady/blob/main/paper/full_paper.md#experiments-and-results)). To briefly describe the approach, we collected actual README files from open-source GitHub repositories and compared them against the README files auto-generated using the `readme_ready` tool. We generated the README by using the same headings as the actual README and calculated the BLEU and BERT scores between them after some minor post-processing to fix the output formatting. The scoring functions can be found under the `Scoring` section in the `scripts/data.ipnb` ([link](https://github.com/souradipp76/ReadMeReady/blob/main/scripts/data.ipynb)). The scoring is done for both non-finetuned and fine-tuned LLAMA models. The finetuned mode can be generated using `scripts/fine-tuning-with-llama-2-qlora.ipynb` notebook.
