### 2024-11-24 4cc34f1
Added mkdocs and minor fixes.
