+++ readme_ready/utils/llm_utils.py
@@ -22,9 +22,7 @@ def get_gemma_chat_model(model_name: str, streaming=False, model_kwargs=None):
         gguf_file = model_kwargs["gguf_file"]
         _ = hf_hub_download(model_name, gguf_file)
     tokenizer = get_tokenizer(model_name, gguf_file)
-    if (
-        sys.platform == "linux" or sys.platform == "linux2"
-    ) and "gptq" not in model_name.lower():
+    if sys.platform != "darwin" and "gptq" not in model_name.lower():
         from transformers import BitsAndBytesConfig
 
         bnb_config = BitsAndBytesConfig(
@@ -57,6 +55,10 @@ def get_gemma_chat_model(model_name: str, streaming=False, model_kwargs=None):
         PEFT_MODEL = model_kwargs["peft_model_path"]
         model = PeftModel.from_pretrained(model, PEFT_MODEL)
 
+    print(
+        f"Memory footprint: {model.get_memory_footprint() / 1024 **3:.2f} GB."
+    )
+
     return HuggingFacePipeline(
         pipeline=pipeline(
             "text-generation",
@@ -78,9 +80,7 @@ def get_llama_chat_model(model_name: str, streaming=False, model_kwargs=None):
         _ = hf_hub_download(model_name, gguf_file)
     tokenizer = get_tokenizer(model_name, gguf_file)
     tokenizer.pad_token = tokenizer.eos_token
-    if (
-        sys.platform == "linux" or sys.platform == "linux2"
-    ) and "gptq" not in model_name.lower():
+    if sys.platform != "darwin" and "gptq" not in model_name.lower():
         from transformers import BitsAndBytesConfig
 
         bnb_config = BitsAndBytesConfig(
@@ -111,6 +111,10 @@ def get_llama_chat_model(model_name: str, streaming=False, model_kwargs=None):
     if "peft_model" in model_kwargs and model_kwargs["peft_model"] is not None:
         PEFT_MODEL = model_kwargs["peft_model"]
         model = PeftModel.from_pretrained(model, PEFT_MODEL)
+
+    print(
+        f"Memory footprint: {model.get_memory_footprint() / 1024 **3:.2f} GB."
+    )
 
     return HuggingFacePipeline(
         pipeline=pipeline(
