+++ docs/index.md
@@ -1,11 +1,123 @@
-# Welcome to ReadMeReady
+# ReadmeReady
 
-## Commands
+Auto-generate code documentation in Markdown format in seconds.
 
-* `readme_ready` - Start generating README documentation.
+## What is ReadmeReady?
 
-## Project layout
+Automated documentation of programming source code is a challenging task with significant practical and scientific implications for the developer community. ReadmeReady is a large language model (LLM)-based application that developers can use as a support tool to generate basic documentation for any publicly available or custom repository. Over the last decade, several research have been done on generating documentation for source code using neural network architectures. With the recent advancements in LLM technology, some open-source applications have been developed to address this problem. However, these applications typically rely on the OpenAI APIs, which incur substantial financial costs, particularly for large repositories. Moreover, none of these open-source applications offer a fine-tuned model or features to enable users to fine-tune custom LLMs. Additionally, finding suitable data for fine-tuning is often challenging. Our application addresses these issues.
 
-    mkdocs.yml    # The configuration file.
-    docs/
-        index.md  # The documentation homepage.
+## Installation
+
+ReadmeReady is available only on Linux/Windows.
+
+### Dependencies
+
+Please follow the installation guide [here](https://pypi.org/project/python-magic/) to install `python-magic`.
+
+### Install it from PyPI
+
+The simplest way to install ReadmeReady and its dependencies is from PyPI with pip, Python's preferred package installer.
+
+```bash
+pip install readme_ready
+```
+
+In order to upgrade ReadmeReady to the latest version, use pip as follows.
+
+```bash
+$ pip install -U readme_ready
+```
+
+### Install it from source
+
+You can also install ReadmeReady from source as follows.
+
+```bash
+$ git clone https://github.com/souradipp76/ReadMeReady.git
+$ cd ReadMeReady
+$ make install
+```
+
+To create a virtual environment before installing ReadmeReady, you can use the command:
+```bash
+$ make virtualenv
+$ source .venv/bin/activate
+```
+
+## Usage
+
+### Initialize
+```bash
+$ export OPENAI_API_KEY=<YOUR_OPENAI_API_KEY>
+$ export HF_TOKEN=<YOUR_HUGGINGFACE_TOKEN>
+```
+
+Set `OPENAI_API_KEY=dummy` to use only open-source models.
+
+### Command-Line
+
+```bash
+$ python -m readme_ready
+#or
+$ readme_ready
+```
+
+### In Code
+
+```py
+from readme_ready.query import query
+from readme_ready.index import index
+from readme_ready.types import (
+    AutodocReadmeConfig,
+    AutodocRepoConfig,
+    AutodocUserConfig,
+    LLMModels,
+)
+
+model = LLMModels.LLAMA2_7B_CHAT_GPTQ # Choose model from supported models
+
+repo_config = AutodocRepoConfig (
+    name = "<NAME>", # Replace <NAME>
+    root = "<PROJECT_ROOT>", # Replace <PROJECT_ROOT>
+    repository_url = "<PROJECT_URL>", # Replace <PROJECT_URL>
+    output = "<OUTPUT_DIR>", # Replace <OUTPUT_DIR>
+    llms = [model],
+    peft_model_path = "<PEFT_MODEL_NAME_OR_PATH>", # Replace <PEFT_MODEL_NAME_OR_PATH>
+    ignore = [
+        ".*",
+        "*package-lock.json",
+        "*package.json",
+        "node_modules",
+        "*dist*",
+        "*build*",
+        "*test*",
+        "*.svg",
+        "*.md",
+        "*.mdx",
+        "*.toml"
+    ],
+    file_prompt = "",
+    folder_prompt = "",
[…]
