+++ docs/reference.md
@@ -1,0 +1,43 @@
+# API Reference
+
+::: readme_ready.index.index
+    handler: python
+    options:
+      members:
+        - index
+      show_root_heading: true
+      show_source: false
+
+::: readme_ready.index.create_vector_store
+    handler: python
+    options:
+      members:
+        - create_vector_store
+      show_root_heading: true
+      show_source: false
+
+::: readme_ready.index.process_repository
+    handler: python
+    options:
+      members:
+        - process_repository
+      show_root_heading: true
+      show_source: false
+      
+::: readme_ready.query.query
+    handler: python
+    options:
+      members:
+        - query
+        - generate_readme
+      show_root_heading: true
+      show_source: false
+
+::: readme_ready.query.create_chat_chain
+    handler: python
+    options:
+      members:
+        - make_qa_chain
+        - make_readme_chain
+      show_root_heading: true
+      show_source: false
