+++ .github/workflows/deploy_mkdocs.yml
@@ -25,7 +25,11 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install mkdocs mkdocs-material mkdocstrings-python
-
+    
+    - name: Setup Pages
+      uses: actions/configure-pages@v3
+    
     # Build the MkDocs site
     - name: Build and Publish MkDocs site
+      uses: actions/deploy-pages@v2
       run: mkdocs gh-deploy
