### 2024-11-24 f2ba81a
Updated pages workflow
