+++ .github/workflows/deploy_mkdocs.yml
@@ -4,6 +4,21 @@ on:
   push:
     branches:
       - main
+  
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+  # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+  permissions:
+    contents: read
+    pages: write
+    id-token: write
+
+  # Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+  # However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+  concurrency:
+    group: "pages"
+    cancel-in-progress: false
 
 jobs:
   deploy:
@@ -13,6 +28,9 @@ jobs:
     # Checkout the repository
     - name: Checkout code
       uses: actions/checkout@v4
+
+    - name: Setup Pages
+        uses: actions/configure-pages@v5
 
     # Set up Python
     - name: Setup Python
@@ -25,9 +43,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install mkdocs mkdocs-material mkdocstrings-python
-    
-    - name: Setup Pages
-      uses: actions/configure-pages@v3
     
     # Build the MkDocs site
     - name: Build and Publish MkDocs site
