+++ .github/workflows/deploy_mkdocs.yml
@@ -30,7 +30,7 @@ jobs:
       uses: actions/checkout@v4
 
     - name: Setup Pages
-        uses: actions/configure-pages@v5
+      uses: actions/configure-pages@v5
 
     # Set up Python
     - name: Setup Python
