+++ .github/workflows/deploy_mkdocs.yml
@@ -8,17 +8,17 @@ on:
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
 
-  # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-  permissions:
-    contents: read
-    pages: write
-    id-token: write
+# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
+permissions:
+  contents: read
+  pages: write
+  id-token: write
 
-  # Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
-  # However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
-  concurrency:
-    group: "pages"
-    cancel-in-progress: false
+# Allow only one concurrent deployment, skipping runs queued between the run in-progress and latest queued.
+# However, do NOT cancel in-progress runs as we want to allow these production deployments to complete.
+concurrency:
+  group: "pages"
+  cancel-in-progress: false
 
 jobs:
   deploy:
