### 2024-11-24 243bf89
Update create_vector_store.py
