+++ readme_ready/query/create_chat_chain.py
@@ -1,5 +1,5 @@
 """
-Create Chat Chain
+Creates Chains for QA Chat or Readme Generation
 """
 
 from typing import List
@@ -12,8 +12,10 @@ from langchain.chains.combine_documents.stuff import (
     create_stuff_documents_chain,
 )
 from langchain.prompts import PromptTemplate
+from langchain_core.runnables.base import Runnable
 
 from readme_ready.types import LLMModels
+from readme_ready.utils.HNSWLib import HNSWLib
 from readme_ready.utils.llm_utils import (
     get_gemma_chat_model,
     get_llama_chat_model,
@@ -139,41 +141,46 @@ def make_readme_prompt(
 
 
 def make_qa_chain(
-    project_name,
-    repository_url,
-    content_type,
-    chat_prompt,
-    target_audience,
-    vectorstore,
+    project_name: str,
+    repository_url: str,
+    content_type: str,
+    chat_prompt: str,
+    target_audience: str,
+    vector_store: HNSWLib,
     llms: List[LLMModels],
     device: str = "cpu",
-    on_token_stream=None,
-):
+    on_token_stream: bool = False,
+) -> Runnable:
     """
-    Creates a question-answering (QA) chain for the specified project.
+    Creates a question-answering (QA) chain for the specified project
 
-    Initializes and configures the QA chain using the provided repository and user configurations.
-    Selects the appropriate language model (LLM), sets up the retriever with a history-aware mechanism,
-    and combines document chains for processing queries. The chain facilitates interaction with the
-    vector store to retrieve and process relevant information based on user queries.
+    Initializes and configures the QA chain using the provided repository
+    and user configurations. Selects the appropriate language model (LLM),
+    sets up the retriever with a history-aware mechanism, and combines
+    document chains for processing queries. The chain facilitates interaction
+    with the vector store to retrieve and process relevant information
+    based on user queries.
 
     Args:
-        project_name: The name of the project for which the QA chain is being created.
+        project_name: The name of the project for which the QA chain is
+            being created.
         repository_url: The URL of the repository containing the project.
-        content_type: The type of content to be processed (e.g., 'code', 'documentation').
+        content_type: The type of content to be processed
+            (e.g., 'code', 'documentation').
         chat_prompt: The prompt template used for generating chat responses.
         target_audience: The intended audience for the QA responses.
-        vectorstore: An instance of HNSWLib representing the vector store containing document embeddings.
-        llms: A list of LLMModels to select from for generating embeddings and responses.
+        vector_store: An instance of HNSWLib representing the vector store
+            containing document embeddings.
+        llms: A list of LLMModels to select from for generating embeddings
+            and responses.
         device: The device to use for model inference (default is 'cpu').
-        on_token_stream: Optional callback for handling token streams during model inference.
+        on_token_stream: Optional callback for handling token streams during
+            model inference.
 
     Returns:
-        A retrieval chain configured for question-answering, combining the retriever and document processing chain.
+        A retrieval chain configured for question-answering, combining the
+            retriever and document processing chain.
 
-    Raises:
-        ValueError: If no suitable model is found in the provided LLMs.
-        RuntimeError: If there is an issue initializing the chat models or creating the chains.
     """
     llm = llms[1] if len(llms) > 1 else llms[0]
     llm_name = llm.value
@@ -206,7 +213,7 @@ def make_qa_chain(
         )
 
     question_generator = create_history_aware_retriever(
-        question_chat_model, vectorstore.as_retriever(), condense_qa_prompt
+        question_chat_model, vector_store.as_retriever(), condense_qa_prompt
     )
 
[…]
