### 2024-11-24 souradipp76/ReadMeReady#12 @Manvi-Agrawal
[Non-Blocking] How about adding `1 × NVIDIA Tesla V100 with 16GB of GPU memory` was used for experimental results in the methodology section?
