### 2024-11-24 souradipp76/ReadMeReady#15 @Manvi-Agrawal
> Project Root refers to the path to the code repository you want to generate a README for. Please download any sample code repository from GitHub and use that or if you have your own repository downloaded (say MyRepo) you can use that as well. Set Project Root as <path to MyRepo> and set the other parameters accordingly. Output Directory is where the README will be generated(you can set any appropriate path). For Windows platforms, please provide the paths correctly (e.g. C:\Users\user\Downloads\MyRepo).

@souradipp76 , thanks for the explaination. Possibly consider adding these explainations in README for future devs.

- Also, could you please explain how can someone use the github url prompt efficiently. I am assuming that would allow users to generate readme for a project existing on github, is that correct?

- Also, I am still getting `FileNotFoundError: Could not read directory: ./readme_ready/.` in google collab with T4 gpu. I set the hf-token and open-ai token as random guids. and ran code in `examples/example.py`
