### 2024-11-25 d02eab1
Updated documentation and default CLI.
