+++ examples/example.py
@@ -10,15 +10,27 @@ from readme_ready.types import (
     LLMModels,
 )
 
-model = LLMModels.LLAMA2_7B_CHAT_GPTQ # Choose model from supported models
+# Choose model from supported models
+model = LLMModels.LLAMA2_7B_CHAT_GPTQ
+
+# Initialize the repository configuration. `root` refers to the path to the
+# code repository for which you want to generate a README for. Please download
+# any code repository from GitHub and use that or if you have your own
+# repository downloaded (say 'MyRepo') you can use that as well.
+# Set `name` to the 'MyRepo'.
+# Set `root` as <path to 'MyRepo'>.
+# Set `repository_url` to the GitHub URL of 'MyRepo' (if any) else leave blank.
+# Set `output` as the path to the directory where the README and other metadata
+# will be generated and saved.
+# Set other parameters accordingly (or leave as default).
 
 repo_config = AutodocRepoConfig (
-    name = "readmy_ready",
-    root = "./readme_ready",
-    repository_url = "https://github.com/souradipp76/ReadMeReady",
-    output = "./output/readmy_ready",
+    name = "readmy_ready", # Set repository name
+    root = "./readme_ready", # Set path to root directory of the repository
+    repository_url = "https://github.com/souradipp76/ReadMeReady", # Set url
+    output = "./output/readmy_ready", # Set path to output directory to save
     llms = [model],
-    peft_model_path = None,
+    peft_model_path = None, # Set path to PEFT model
     ignore = [
         ".*",
         "*package-lock.json",
@@ -41,7 +53,7 @@ repo_config = AutodocRepoConfig (
     priority = None,
     max_concurrent_calls = 50,
     add_questions = False,
-    device = "auto",
+    device = "auto", # Select device "cpu" or "auto"
 )
 
 user_config = AutodocUserConfig(
@@ -49,6 +61,7 @@ user_config = AutodocUserConfig(
 )
 
 readme_config = AutodocReadmeConfig(
+    # Set comma separated list of README headings
     headings = "Description,Requirements,Installation,Usage,Contributing,License"
 )
 
