### 2024-11-25 souradipp76/ReadMeReady#15 @souradipp76
@Manvi-Agrawal, I have added additional instructions in README as well in `examples/example.py`.
- In case there is no GitHub URL, the user can leave it blank.
- For the `FileNotFoundError`, please set the exact directory path where the repository is located. The paths need to be changed accordingly depending on which directory you are running the script from.
