### 2024-11-25 souradipp76/ReadMeReady#13 @souradipp76
Added the link to API docs on the README.
