### 2024-12-24 #7489 @Manvi-Agrawal
Hi @crvernon, I am hoping to get to review this work this week or next. I would greatly appreciate if the steps for artifact evaluation via Google collab GPU could be added. This would greatly reduce effort on my side and improve onboarding experience for the new users. It has been sometime since I used collab for data driven experiments.

I referenced a link earlier regarding a sample artifact. I need some time to find another link that has very simple steps to reproduce the results. I think we could add some examples to the JOSS website too.

PS : sent from mobile. Not responsible for autocorrect errors
