### 2025-01-03 e4d39a6
Added notebook example.
