### 2025-01-05 #7489 @souradipp76
@Manvi-Agrawal, I have created an example notebook which you can run in Google Colab with GPU ([here](https://colab.research.google.com/github/souradipp76/ReadMeReady/blob/main/examples/example.ipynb)) for validating the tool usage. Let me know if you are still facing any issue on getting started with the tool. Also we are working on making it easier to reproduce the results in the paper. Will keep you updated.
