### 2025-01-05 souradipp76/ReadMeReady#15 @souradipp76
@Manvi-Agrawal Please see the example to run in Google Colab with GPU as mentioned [here](https://github.com/openjournals/joss-reviews/issues/7489#issuecomment-2571711162).
