### 2025-01-05 souradipp76/ReadMeReady#15 @Manvi-Agrawal
Thanks @souradipp76 the shared example worked without any issues with T4 GPU. I ended up commenting `login(token=HF_TOKEN, add_to_git_credential=False)` and used fake HF token for `LLAMA2_7B_CHAT_GPTQ` model.

Closing the issue since I am able to run on GPU.
