### 2025-01-20 81ea9c1
Added script to run validations.
