+++ scripts/validate.py
@@ -1,0 +1,224 @@
+import re
+import bert_score
+from nltk.translate.bleu_score import sentence_bleu, SmoothingFunction
+from transformers import AutoTokenizer
+
+import os
+import subprocess
+import pandas as pd
+
+from readme_ready.query import query
+from readme_ready.index import index
+from readme_ready.types import (
+    AutodocReadmeConfig,
+    AutodocRepoConfig,
+    AutodocUserConfig,
+    LLMModels,
+)
+
+# Choose model from supported models
+model = LLMModels.LLAMA2_7B_CHAT_GPTQ
+
+# Clone repository to a local path
+def git_clone(repo_url, clone_path):
+    if os.path.exists(clone_path):
+        subprocess.run(['rm', '-rf', clone_path], check=True)
+    subprocess.run(['git', 'clone', repo_url, clone_path], check=True)
+    subprocess.run(['rm', '-rf', os.path.join(clone_path,".git")], check=True)
+    subprocess.run(['rm', '-rf', os.path.join(clone_path,".gitignore")], check=True)
+
+# Parse the README.md content
+def parse_markdown(md_file_path):
+    with open(md_file_path, 'r', encoding='utf-8') as file:
+        md_content = file.read()
+
+    heading_pattern = re.compile(r'^(#+)\s*(.*)', re.MULTILINE)
+    headings_contents = []
+    current_heading = None
+    current_content = []
+
+    for line in md_content.split('\n'):
+        match = heading_pattern.match(line)
+        if match:
+            if current_heading is not None:
+                headings_contents.append([current_heading, ' '.join(current_content).strip()])
+            current_heading = match.group(2).strip()
+            current_content = []
+        else:
+            if line.strip():
+                current_content.append(line.strip())
+
+    if current_heading is not None:
+        headings_contents.append([current_heading, ' '.join(current_content).strip()])
+
+    df = pd.DataFrame(headings_contents, columns=['Title', 'Content'])
+    return df, md_content
+
+def clean_emoji(tx):
+    emoji_pattern = re.compile("["
+        u"\U0001F600-\U0001F64F"  # emoticons
+        u"\U0001F300-\U0001F5FF"  # symbols
+        u"\U0001F680-\U0001F6FF"  # transport
+        u"\U0001F1E0-\U0001F1FF"  # flags
+        u"\U00002702-\U000027B0"
+        u"\U000024C2-\U0001F251"
+        "]+", flags=re.UNICODE
+    )
+
+    return emoji_pattern.sub(r'', tx)
+
+def text_cleaner(tx):
+    text = re.sub(r"won\'t", "would not", tx)
+    text = re.sub(r"im", "i am", tx)
+    text = re.sub(r"Im", "I am", tx)
+    text = re.sub(r"can\'t", "can not", text)
+    text = re.sub(r"don\'t", "do not", text)
+    text = re.sub(r"shouldn\'t", "should not", text)
+    text = re.sub(r"needn\'t", "need not", text)
+    text = re.sub(r"hasn\'t", "has not", text)
+    text = re.sub(r"haven\'t", "have not", text)
+    text = re.sub(r"weren\'t", "were not", text)
+    text = re.sub(r"mightn\'t", "might not", text)
+    text = re.sub(r"didn\'t", "did not", text)
+    text = re.sub(r"n\'t", " not", text)
+    text = re.sub(r"\'re", " are", text)
+    text = re.sub(r"\'s", " is", text)
+    text = re.sub(r"\'d", " would", text)
+    text = re.sub(r"\'ll", " will", text)
+    text = re.sub(r"\'t", " not", text)
+    text = re.sub(r"\'ve", " have", text)
[…]
