### 2025-01-21 a1a4cdc
Minor fix in validation script.
