### 2025-01-21 4a087c1
Updated validation script.
