### 2025-01-22 souradipp76/ReadMeReady#11 @souradipp76
Hi @Manvi-Agrawal, I have added a script scripts/run_validate.sh([link](https://github.com/souradipp76/ReadMeReady/blob/main/scripts/run_validate.sh)) that you can run to reproduce the BLEU and BERT scores for the 5 sample repositories shown in the paper for the W/O Finetuning case using quantized LLAMA2 model without any manual post-processing(to fix the formatting). However, it is to be noted that running the script requires atleast 16GB of GPU and may take a long time without GPU.
