### 2025-01-23 souradipp76/ReadMeReady#11 @Manvi-Agrawal
Thanks @souradipp76 for adding the script. For easier and faster validation, could you please create a notebook, so that I can verify via collab? Thank you and looking forward for the review.
