### 2025-01-25 95d4e7d
Add notebook for validation.
