### 2025-01-25 souradipp76/ReadMeReady#11 @souradipp76
@Manvi-Agrawal  This is the notebook ([link](https://colab.research.google.com/github/souradipp76/ReadMeReady/blob/main/scripts/validate.ipynb)) that you ca run in Google Colab.
