### 2025-02-14 souradipp76/ReadMeReady#11 @souradipp76
Hi @Manvi-Agrawal, any updates on the review of this issue? I think you intended to do some performance testing as mentioned [here](https://github.com/souradipp76/ReadMeReady/issues/17#issuecomment-2635863682). However, our paper does not have any performance claims and only aims to provide this as an experimental software tool. The notebook to reproduce the results have been provided above. Let me know if you are able to run the notebook.
