### 2025-02-15 souradipp76/ReadMeReady#11 @Manvi-Agrawal
Hi @souradipp76, I tried to reproduce the scores mentioned in `paper.md` using the python notebook that was shared using T4 gpu. However, the execution timed out twice. I did verify earlier that generating READMEs work.

 I am not sure if verifying the scores is under the purview of my responsibilities. If so, do you mind trying the verification on your system using **google collab T4 free version**? I have heard that checkpointing results by mounting drive helps in these cases, but I havent gotten time to try it out. @crvernon, @camilochs, what do you think of this ?
