### 2025-02-15 souradipp76/ReadMeReady#11 @camilochs
@Manvi-Agrawal, I tried to replicate the score values but it keeps computing in the function: `generate_readme(name, url, repo_root, output, headings)`. I even chose only a small repository and the same thing happens. Maybe it is something minor, you could check it @souradipp76.
