### 2025-02-15 souradipp76/ReadMeReady#11 @Manvi-Agrawal
@camilochs , running [example.ipynb](https://github.com/souradipp76/ReadMeReady/blob/main/examples/example.ipynb) worked for me which was trying to create readme for Touchpose repository. I commented the login line during the API token setup cell.

I am facing `generate_readme` step using T4 GPU,  as you rightly mentioned, when running "validate.ipynb". I suspect the issue is arising due to GPU timeout in collab free version which I expect to take roughly 5 hours (10 hours, each of ~30 min). @souradipp76 could you please help in scores reproducibility?
