### 2025-02-21 souradipp76/ReadMeReady#11 @souradipp76
@Manvi-Agrawal @camilochs, it would be difficult to run the reproducibility script in the free version of Google Colab as it frequently times out. We ran the script in a server containing 1 × NVIDIA Tesla V100 with 16GB of GPU memory. One suggestion that I can give is to run the script for 1 repository at a time. You can comment out all other repositories from the list of repositories except one in the `validation.ipynb` and run the script. I hope that helps. I am also looking into ways to make the processing faster but working with LLMs like LLAMA poses certain challenges especially when we are constrained by hardware resources.
