### 2025-03-16 #7489 @hellokayas
> Hi [@crvernon](https://github.com/crvernon) [@camilochs](https://github.com/camilochs) , I think we have done primary validation from our side.The only thing that remains is validation of the scores. since the focus of this project is to generate readmes, I was thinking if it might be okay to skip score validation, because it would mean investing a few hours from both me and [@camilochs](https://github.com/camilochs), because we lack GPU access and would need to use Google collab free version.
>
> Instead we could ask Souradip to remove the scores part from the paper and add a note about different scores and optimization as a footnote in the documentation. [@crvernon](https://github.com/crvernon) [@camilochs](https://github.com/camilochs) thoughts

@camilochs @crvernon @Manvi-Agrawal We would be happy to do that, no worries. We will do that within this week. Please go ahead with the other check boxes.
