### 2025-03-18 #7489 @hellokayas
@camilochs @Manvi-Agrawal We have removed the scores part from the paper and added that section about different scores and optimization in the documentation readme. Kindly proceed with the review. Let us know if any more changes are required.
