+++ paper/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @article{miller2003cognitive,
-  title={The cognitive revolution: a historical perspective},
+  title={The cognitive revolution: {a} historical perspective},
   author={Miller, George A},
   journal={Trends in cognitive sciences},
   volume={7},
@@ -25,7 +25,7 @@
 }
 
 @article{graves2014neural,
-  title={Neural turing machines},
+  title={Neural {T}uring {M}achines},
   author={Graves, Alex and Wayne, Greg and Danihelka, Ivo},
   journal={arXiv preprint arXiv:1410.5401},
   year={2014},
@@ -33,7 +33,7 @@
 }
 
 @article{bahdanau2014neural,
-  title={Neural machine translation by jointly learning to align and translate},
+  title={Neural Machine Translation by Jointly Learning to Align and Translate},
   author={Bahdanau, Dzmitry and Cho, Kyunghyun and Bengio, Yoshua},
   journal={arXiv preprint arXiv:1409.0473},
   year={2014},
@@ -41,7 +41,7 @@
 }
 
 @inproceedings{iyer2016summarizing,
-  title={Summarizing source code using a neural attention model},
+  title={Summarizing Source Code using a Neural Attention Model},
   author={Iyer, Srinivasan and Konstas, Ioannis and Cheung, Alvin and Zettlemoyer, Luke},
   booktitle={54th Annual Meeting of the Association for Computational Linguistics 2016},
   pages={2073--2083},
@@ -51,7 +51,7 @@
 }
 
 @article{ling2016latent,
-  title={Latent predictor networks for code generation},
+  title={Latent Predictor Networks for Code Generation},
   author={Ling, Wang and Grefenstette, Edward and Hermann, Karl Moritz and Ko{\v{c}}isk{\`y}, Tom{\'a}{\v{s}} and Senior, Andrew and Wang, Fumin and Blunsom, Phil},
   journal={arXiv preprint arXiv:1603.06744},
   year={2016},
@@ -59,7 +59,7 @@
 }
 
 @article{yin2017syntactic,
-  title={A syntactic neural model for general-purpose code generation},
+  title={A Syntactic Neural Model for General-purpose Code Generation},
   author={Yin, Pengcheng and Neubig, Graham},
   journal={arXiv preprint arXiv:1704.01696},
   year={2017},
@@ -77,7 +77,7 @@
 }
 
 @article{bhoopchand2016learning,
-  title={Learning python code suggestion with a sparse pointer network},
+  title={Learning Python Code Suggestion with a Sparse Pointer Network},
   author={Bhoopchand, Avishkar and Rockt{\"a}schel, Tim and Barr, Earl and Riedel, Sebastian},
   journal={arXiv preprint arXiv:1611.08307},
   year={2016},
@@ -85,7 +85,7 @@
 }
 
 @inproceedings{oda2015learning,
-  title={Learning to generate pseudo-code from source code using statistical machine translation},
+  title={Learning to Generate Pseudo-Code from Source Code using Statistical Machine Translation},
   author={Oda, Yusuke and Fudaba, Hiroyuki and Neubig, Graham and Hata, Hideaki and Sakti, Sakriani and Toda, Tomoki and Nakamura, Satoshi},
   booktitle={2015 30th IEEE/ACM International Conference on Automated Software Engineering (ASE)},
   pages={574--584},
@@ -104,7 +104,7 @@
 }
 
 @article{radford2019language,
-  title={Language models are unsupervised multitask learners},
+  title={Language {M}odels are {U}nsupervised {M}ultitask {L}earners},
[…]
