### 2025-04-12 #7489 @hellokayas
> 👋 [@hellokayas](https://github.com/hellokayas) - we are almost there!
>
> Here are a few things that need to be addressed in your paper:
>
> * [ ]  LINE 23:  "...reasoning((Chomsky, 1956), (Miller, 2003), (Graves et al., 2014))."  There is a missing space in between "reasoning" and your first parenthesis.  Also, the citations here should appear as "...reasoning (Chomsky, 1956; Miller, 2003; Graves et al., 2014)."  In you paper.md file this is currently represented as "reasoning([@chomsky1956three], [@miller2003cognitive], [@graves2014neural])." but can be fixed using "reasoning [@chomsky1956three; @miller2003cognitive; @graves2014neural]."[ ]  LINE 26-27:  Same formatting issue for multiple citations as above.[ ]  LINE 42:  Missing space between "QLoRA" and the citation.[ ]  LINE 139:  The "t" in "turing" should be capitalized.  You can do this in your bib file by adding curly brackets around characters you wish to maintain the formatting of.
>
> Next is just setting up the archive for your new release.
>
> We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.
>
> So here is what we have left to do:
>
> * [ ]
>   [x]   Confirm that the title of your paper in your paper.md file has no special formatting (e.g., backticks) as this will cause formatting issues when accepted.
>   [ ]
>   [x]   Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository).  Please ensure that the software archive uses the same license as the license you have posted on code repository.
>   [ ]
>   [x]   Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).
>   [ ]
>   [x]   Please respond with the DOI of the archived version here along with the version number released.
>
> I can then move forward with accepting the submission.

@crvernon All the comments about the lines in the paper have been fixed.

The DOI of the archived version is 10.5281/zenodo.15201191
The version number is 1.1.5

Please go ahead with the publication process.
