### 2025-04-12 #7489 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
