### 2025-04-12 #7489 @crvernon
🥳 Congratulations on your new publication @souradipp76 and @hellokayas! Many thanks to @Manvi-Agrawal and @camilochs for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
