### 2025-05-14 #7489 @ritika-giri
Hi @souradipp76 - thanks for checking in.
Yes, it is OK to submit this work for a talk or one of the other submission formats like poster, notebook, etc **other than the paper format** since this is already published as a paper in JOSS. Please let me know if you have other questions!
