### 2025-05-14 #7489 @hellokayas
> Thanks for checking in. Yes, it is OK to submit this work for a talk or one of the other submission formats like poster, notebook, etc **other than the paper format** since this is already published as a paper in JOSS. Please let me know if you have other questions!

@ritika-giri That sounds like a good idea. I think submitting our work as a poster or talk is something we can do. Do you know if the travel and accommodation are covered in case our work gets selected? Let us know the info you have about submitting as a talk/poster.
