### 2025-05-14 #7489 @ritika-giri
Hi @hellokayas -  yes absolutely, you can submit for a talk or poster - or even as a workshop if you think that presenting it as a 90 minute demo would be better. Submission types and details are [here](https://us-rse.org/usrse25/participate/#submission-types). I'm also happy to answer more Qs or connect you to the technical program chairs if you'd like.
Unfortunately we don't currently have any travel support. If we do, it will be limited to authors based in the US, and we will keep folks informed of funding options via email.
Last year we had a lot of interest in using LLMs for code testing and documentation, so your work will be well-received by the community! Also, please feel free to [join](https://docs.google.com/forms/d/e/1FAIpQLSdJbPczGHFN8mfMFu_YQym508OzFtOZxfSzr1sOoINxaMmiaw/viewform) the US-RSE Slack if you'd like to connect with the community.
