### 2025-05-15 #7489 @hellokayas
@ritika-giri Thanks so much for the details. I think we will go for the talk or poster and **not** for the 90-minute demo. Also, it will be very helpful if you can put us in touch with the technical program chairs.
About the funding, I wanted to let you know that we are both based in the US. So if any funding opportunity opens up, let us know.
