### 2025-05-15 #7489 @hellokayas
> Hi [@hellokayas](https://github.com/hellokayas) - yes absolutely, you can submit for a talk or poster - or even as a workshop if you think that presenting it as a 90 minute demo would be better. Submission types and details are [here](https://us-rse.org/usrse25/participate/#submission-types). I'm also happy to answer more Qs or connect you to the technical program chairs if you'd like. Unfortunately we don't currently have any travel support. If we do, it will be limited to authors based in the US, and we will keep folks informed of funding options via email. Last year we had a lot of interest in using LLMs for code testing and documentation, so your work will be well-received by the community! Also, please feel free to [join](https://docs.google.com/forms/d/e/1FAIpQLSdJbPczGHFN8mfMFu_YQym508OzFtOZxfSzr1sOoINxaMmiaw/viewform) the US-RSE Slack if you'd like to connect with the community.

@ritika-giri Thanks so much for the details. I think we will go for the talk or poster and not for the 90-minute demo. Also, it will be very helpful if you can put us in touch with the technical program chairs.
About the funding, I wanted to let you know that we are both based in the US. So if any funding opportunity opens up, let us know.
